Fixing rotational sweep structure fields
Adding proper handling for body frame to model frame coordinate transformations.

Challenge is that structure fields are logically defined in the model space of the geoemtry. However, the simualtor store geometries in body frame coordinate system. Hence one must take this transform into account when generating structure fields in the body frames of the rigid bodies.